requirements.txt
streamlit==1.28.1
fastapi==0.100.0
uvicorn==0.23.2
transformers==4.35.2
torch==2.9.0
openai-whisper==20231117
pydantic==2.5.0
langchain==0.0.350
selenium==4.15.2
webdriver-manager==4.0.1
pillow==10.1.0